feat: fix compatibility with go 1.18
`Unwrap() []error` is only supported in the `errors` package
starting go 1.20, so attempting to test a `ContentError`
(which implements `Unwrap() []error`) with `errors.Is(...)`
will not work.

This commit pulls that specific test into a separate file only
compiled with go 1.20 and tests the Unwrap methods directly
to ensure functionality on go 1.18.

Fixes #5